requirement.txt
1.Flask
2.randome